INTERNAL: Move exception message logic into each exception class
