+++ src/main/java/net/spy/memcached/internal/CheckedOperationTimeoutException.java
@@ -19,11 +19,14 @@ package net.spy.memcached.internal;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import net.spy.memcached.ExceptionMessageFactory;
+import net.spy.memcached.MemcachedNode;
+import net.spy.memcached.ops.DeleteOperation;
 import net.spy.memcached.ops.Operation;
+import net.spy.memcached.ops.StoreOperation;
 
 /**
  * Timeout exception that tracks the original operation.
@@ -44,14 +47,71 @@ public class CheckedOperationTimeoutException extends TimeoutException {
                                           TimeUnit unit,
                                           long elapsed,
                                           Collection<Operation> ops) {
-    super(ExceptionMessageFactory.createTimedOutMessage(duration, unit, elapsed, ops));
+    super(createMessage(duration, unit, elapsed, ops));
     operations =  new ArrayList<>(ops);
   }
 
   /**
    * Get the operation that timed out.
    */
-  public Collection<Operation> getOperations() {
+  public List<Operation> getOperations() {
     return operations;
+  }
+
+  private static String createMessage(long duration,
+                                     TimeUnit unit,
+                                     long elapsed,
+                                     Collection<Operation> ops) {
+
+    StringBuilder rv = new StringBuilder();
+    Operation firstOp = ops.iterator().next();
+    if (isBulkOperation(firstOp, ops)) {
+      rv.append("bulk ");
+    }
+    if (firstOp.isPipeOperation()) {
+      rv.append("pipe ");
+    }
+
+    rv.append(firstOp.getAPIType());
+    rv.append(" operation timed out (");
+    rv.append(unit.convert(elapsed, TimeUnit.MILLISECONDS));
+    rv.append(" >= ").append(duration);
+    rv.append(" ").append(unit).append(") - ");
+
+    rv.append("failing node");
+    rv.append(ops.size() == 1 ? ": " : "s: ");
+
+    boolean first = true;
+    for (Operation op : ops) {
+      if (first) {
+        first = false;
+      } else {
+        rv.append(", ");
+      }
+      MemcachedNode node = op == null ? null : op.getHandlingNode();
+      rv.append(node == null ? "<unknown>" : node.getNodeName());
+      if (op != null) {
+        rv.append(" [").append(op.getState()).append("]");
+      }
+      if (node != null) {
+        rv.append(" [").append(node.getOpQueueStatus()).append("]");
+        if (!node.isActive() && node.isFirstConnecting()) {
+          rv.append(" (Not connected yet)");
+        }
+      }
+    }
+    return rv.toString();
+  }
+
+  /**
+   * check bulk operation or not
+   * @param op operation
+   * @param ops number of operation (used in StoreOperationImpl, DeleteOperationImpl)
+   */
+  private static boolean isBulkOperation(Operation op, Collection<Operation> ops) {
+    if (op instanceof StoreOperation || op instanceof DeleteOperation) {
+      return ops.size() > 1;
+    }
+    return op.isBulkOperation();
   }
 }
